Create Toy_modelv2
This model will run once and later will ask you to input a word. If the word is in the dictionary of words, it will try to construct a sentences of size: 5 (see, "size_of_sentence=5" line 161)
It will continue asking your words until you type the word: ESC
I modified the line 39 from:  separate = part_S.split(" ") to:  separate = part_S.split() that seemed to have fixed the appearance of the symbol: \n.
Ask check up it prints: 
1)        line 40:         print('separate words:',separate). Shows the sentences properly tokenised.
2)       line 58:        print("\n full vocabulary: \n", vocabulary) Shows the individual  words that form the vocabulary
3)       line 64:           print("dictionary of words: \n", word2idx). Print the dictionary created associating word to index
4)       line 65:       print("number of words in the dictionary", vocabulary_size).  Print the number of words in the dictionary

New feature: created a function: sentence_production(first_word) line 152
it tries to predict the next word given the input word and create a chain to produce a sentence. 
It is not perfect and we need to understand better skiagram  model to make better predictions. At the moment the sentences are quite gibberish and the probability of the next word is not taking into account the previous words already printed.